+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
     * Maker Checker 
         * [CP-3873] - KYC Maker Checker UI Improvements
         * [CP-3911] - Apply KYC backend changes to align with the updated requirements.
+        * [CP-3927] - Enable search by client name and order by date on Failed KYC tab.
 ## Version 1.4.1 - Community 1.0.0
 
     * Clients